Added prototype functionality for performing region of interest (ROI)
mapping for large molecules. This allows for the set up of protein FEP calculations (for example point mutations and covalent modifications).  The merge code has also been modified to allow for multiple perturbable region of interests, which allows for multiple mutations FEP calculations to be run at the same time.