Correct restraint search torsional variance calculation
Previously, the variance of torsions was being calculated as ~3 times
greater than it should have been. Now, the selected torsional force
constants will be ~3 times lower, and slightly different anchor points
may be selected as the torsions will have slightly less weight when
ordering candidate anchor points. However, this does not affect the
correctness of any previous ABFE calculations, and stability should not
be affected (as this is related to the angle restraints, and not the
torsional restraints).